requirements.txt
mysql-connector-python>=8.3.0
requests>=2.31.0
mutagen>=1.47.0
nest-asyncio
httpx>=0.26.0
gTTS>=2.5.1
python-dateutil>=2.8.2
discord.py>=2.3.2
discord.py[voice]
levenshtein
matplotlib
fastapi
uvicorn
sqlalchemy
python-dotenv